Decimal

+++ src/main/java/com/jhw/gestion/modules/gasto/core/usecase_def/TipoGastoUseCase.java
@@ -2,9 +2,10 @@ package com.jhw.gestion.modules.gasto.core.usecase_def;
 
 import com.clean.core.app.usecase.CRUDUseCase;
 import com.jhw.gestion.modules.gasto.core.domain.*;
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 public interface TipoGastoUseCase extends CRUDUseCase<TipoGastoDomain> {
 
-    public HashMap<TipoGastoDomain, Double> reportGastadoPorGasto() throws Exception;
+    public HashMap<TipoGastoDomain, BigDecimal> reportGastadoPorGasto() throws Exception;
 }
